+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- feat: add empty compute plan status (#190)
+
 ## [0.22.0](<https://github.com/owkin/substra/releases/tag/0.22.0>) - 2022-05-22
 
 ### Added